修正

+++ src/math/select_reduce_max_d_sub_exp_reduce_sum_d_real_div/examples/AclNNInvocationNaive/gen_data.py
@@ -32,7 +32,7 @@ def gen_golden_data_simple():
 
     os.system("mkdir -p input")
     os.system("mkdir -p output")
-    sel.astype(dtype).tofile("./input/input_1.bin")
+    sel.tofile("./input/input_1.bin")
     input1.astype(dtype).tofile("./input/input_2.bin")
     input2.astype(dtype).tofile("./input/input_3.bin")
     result.tofile("./output/golden.bin")
